Fixed copy-paste error in last commit. JIRA:CXX-11310
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@89659 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/ncbistr.cpp
@@ -1893,7 +1893,7 @@ void NStr::LongToString(string& out_str, long svalue,
         return;
     }
     if ( base == 10  ) {
-        unsigned int value = static_cast<unsigned int>(svalue<0?-svalue:svalue);
+        unsigned long value = static_cast<unsigned long>(svalue<0?-svalue:svalue);
         s_SignedBase10ToString(out_str, value, svalue, flags, base);
     } else {
         s_UnsignedOtherBaseToString(out_str, static_cast<unsigned long>(svalue), flags, base);
